Handle the case of no claims.
I had to make one change to handle the case when there are no claims.
Which was an odd case. I checked the few times it occurred and it was
when the the XML string wasn't for an application, but for what looked
like a DNA sequence. It had a DOCTYPE that said sequence-listing and the
data that followed wasn't obviously not an application. In the end it
doesn't do anything to the DB for those, I just had to change the code
to handle the case of the regex for claims not finding any match.